+++ crates/bevy_render/src/render_phase/mod.rs
@@ -1,14 +1,55 @@
+//! The modular rendering abstraction responsible for queuing, preparing, sorting and drawing
+//! entities as part of separate render phases.
+//!
+//! In Bevy each view (camera, or shadow-casting light, etc.) has one or multiple [`RenderPhase`]s
+//! (e.g. opaque, transparent, shadow, etc).
+//! They are used to queue entities for rendering.
+//! Multiple phases might be required due to different sorting/batching behaviours
+//! (e.g. opaque: front to back, transparent: back to front) or because one phase depends on
+//! the rendered texture of the previous phase (e.g. for screen-space reflections).
+//!
+//! To draw an entity, a corresponding [`PhaseItem`] has to be added to one or multiple of these
+//! render phases for each view that it is visible in.
+//! This must be done in the [`RenderStage::Queue`](crate::RenderStage::Queue).
+//! After that the render phase sorts them in the
+//! [`RenderStage::PhaseSort`](crate::RenderStage::PhaseSort).
+//! Finally the items are rendered using a single [`TrackedRenderPass`], during the
+//! [`RenderStage::Render`](crate::RenderStage::Render).
+//!
+//! Therefore each phase item is assigned a [`Draw`] function.
+//! These set up the state of the [`TrackedRenderPass`] (i.e. select the
+//! [`RenderPipeline`](crate::render_resource::RenderPipeline), configure the
+//! [`BindGroup`](crate::render_resource::BindGroup)s, etc.) and then issue a draw call,
+//! for the corresponding item.
+//!
+//! The [`Draw`] function trait can either be implemented directly or such a function can be
+//! created by composing multiple [`RenderCommand`]s.
+
 mod draw;
 mod draw_state;
+mod rangefinder;
 
-use bevy_ecs::entity::Entity;
 pub use draw::*;
 pub use draw_state::*;
+pub use rangefinder::*;
 
-use bevy_ecs::prelude::{Component, Query};
-use bevy_ecs::world::World;
+use crate::render_resource::{CachedRenderPipelineId, PipelineCache};
+use bevy_ecs::{
+    prelude::*,
+    system::{lifetimeless::SRes, SystemParamItem},
+};
+use std::ops::Range;
 
-/// A resource to collect and sort draw requests for specific [`PhaseItems`](PhaseItem).
+/// A collection of all rendering instructions, that will be executed by the GPU, for a
+/// single render phase for a single view.
+///
+/// Each view (camera, or shadow-casting light, etc.) can have one or multiple render phases.
+/// They are used to queue entities for rendering.
+/// Multiple phases might be required due to different sorting/batching behaviours
+/// (e.g. opaque: front to back, transparent: back to front) or because one phase depends on
+/// the rendered texture of the previous phase (e.g. for screen-space reflections).
+/// All [`PhaseItem`]s are then rendered using a single [`TrackedRenderPass`].
+/// The render pass might be reused for multiple phases to reduce GPU overhead.
 #[derive(Component)]
 pub struct RenderPhase<I: PhaseItem> {
     pub items: Vec<I>,
@@ -27,11 +68,12 @@ impl<I: PhaseItem> RenderPhase<I> {
         self.items.push(item);
     }
 
-    /// Sorts all of its [`PhaseItems`](PhaseItem).
+    /// Sorts all of its [`PhaseItem`]s.
     pub fn sort(&mut self) {
         I::sort(&mut self.items);
     }
 
+    /// Renders all of its [`PhaseItem`]s using their corresponding draw functions.
     pub fn render<'w>(
         &self,
         render_pass: &mut TrackedRenderPass<'w>,
@@ -76,7 +118,138 @@ impl<I: BatchedPhaseItem> RenderPhase<I> {
     }
 }
 
-/// This system sorts all [`RenderPhases`](RenderPhase) for the [`PhaseItem`] type.
+/// An item (entity of the render world) which will be drawn to a texture or the screen,
+/// as part of a [`RenderPhase`].
+///
+/// The data required for rendering an entity is extracted from the main world in the
+/// [`RenderStage::Extract`](crate::RenderStage::Extract).
+/// Then it has to be queued up for rendering during the
+/// [`RenderStage::Queue`](crate::RenderStage::Queue), by adding a corresponding phase item to
+/// a render phase.
+/// Afterwards it will be sorted and rendered automatically in the
+/// [`RenderStage::PhaseSort`](crate::RenderStage::PhaseSort) and
+/// [`RenderStage::Render`](crate::RenderStage::Render), respectively.
+pub trait PhaseItem: Sized + Send + Sync + 'static {
+    /// The type used for ordering the items. The smallest values are drawn first.
+    /// This order can be calculated using the [`ViewRangefinder3d`],
+    /// based on the view-space `Z` value of the corresponding view matrix.
+    type SortKey: Ord;
+
+    /// The corresponding entity that will be drawn.
+    ///
+    /// This is used to fetch the render data of the entity, required by the draw function,
+    /// from the render world .
+    fn entity(&self) -> Entity;
+
+    /// Determines the order in which the items are drawn.
+    fn sort_key(&self) -> Self::SortKey;
+
+    /// Specifies the [`Draw`] function used to render the item.
+    fn draw_function(&self) -> DrawFunctionId;
+
+    /// Sorts a slice of phase items into render order. Generally if the same type
+    /// implements [`BatchedPhaseItem`], this should use a stable sort like [`slice::sort_by_key`].
+    /// In almost all other cases, this should not be altered from the default,
+    /// which uses a unstable sort, as this provides the best balance of CPU and GPU
+    /// performance.
+    ///
+    /// Implementers can optionally not sort the list at all. This is generally advisable if and
+    /// only if the renderer supports a depth prepass, which is by default not supported by
+    /// the rest of Bevy's first party rendering crates. Even then, this may have a negative
+    /// impact on GPU-side performance due to overdraw.
+    ///
+    /// It's advised to always profile for performance changes when changing this implementation.
+    #[inline]
+    fn sort(items: &mut [Self]) {
+        items.sort_unstable_by_key(|item| item.sort_key());
+    }
+}
+
+/// A [`PhaseItem`] item, that automatically sets the appropriate render pipeline,
+/// cached in the [`PipelineCache`].
+///
+/// You can use the [`SetItemPipeline`] render command to set the pipeline for this item.
+pub trait CachedRenderPipelinePhaseItem: PhaseItem {
+    /// The id of the render pipeline, cached in the [`PipelineCache`], that will be used to draw
+    /// this phase item.
+    fn cached_pipeline(&self) -> CachedRenderPipelineId;
+}
+
+/// A [`RenderCommand`] that sets the pipeline for the [`CachedRenderPipelinePhaseItem`].
+pub struct SetItemPipeline;
+
+impl<P: CachedRenderPipelinePhaseItem> RenderCommand<P> for SetItemPipeline {
+    type Param = SRes<PipelineCache>;
+    type ViewWorldQuery = ();
+    type ItemWorldQuery = ();
+    #[inline]
+    fn render<'w>(
+        item: &P,
+        _view: (),
+        _entity: (),
+        pipeline_cache: SystemParamItem<'w, '_, Self::Param>,
+        pass: &mut TrackedRenderPass<'w>,
+    ) -> RenderCommandResult {
+        if let Some(pipeline) = pipeline_cache
+            .into_inner()
+            .get_render_pipeline(item.cached_pipeline())
+        {
+            pass.set_render_pipeline(pipeline);
+            RenderCommandResult::Success
+        } else {
+            RenderCommandResult::Failure
+        }
+    }
+}
+
+/// A [`PhaseItem`] that can be batched dynamically.
+///
+/// Batching is an optimization that regroups multiple items in the same vertex buffer
+/// to render them in a single draw call.
+///
+/// If this is implemented on a type, the implementation of [`PhaseItem::sort`] should
+/// be changed to implement a stable sort, or incorrect/suboptimal batching may result.
+pub trait BatchedPhaseItem: PhaseItem {
+    /// Range in the vertex buffer of this item.
+    fn batch_range(&self) -> &Option<Range<u32>>;
+
+    /// Range in the vertex buffer of this item.
+    fn batch_range_mut(&mut self) -> &mut Option<Range<u32>>;
+
+    /// Batches another item within this item if they are compatible.
+    /// Items can be batched together if they have the same entity, and consecutive ranges.
+    /// If batching is successful, the `other` item should be discarded from the render pass.
+    #[inline]
+    fn add_to_batch(&mut self, other: &Self) -> BatchResult {
+        let self_entity = self.entity();
+        if let (Some(self_batch_range), Some(other_batch_range)) = (
+            self.batch_range_mut().as_mut(),
+            other.batch_range().as_ref(),
+        ) {
+            // If the items are compatible, join their range into `self`
+            if self_entity == other.entity() {
+                if self_batch_range.end == other_batch_range.start {
+                    self_batch_range.end = other_batch_range.end;
+                    return BatchResult::Success;
+                } else if self_batch_range.start == other_batch_range.end {
+                    self_batch_range.start = other_batch_range.start;
+                    return BatchResult::Success;
+                }
+            }
+        }
+        BatchResult::IncompatibleItems
+    }
+}
+
+/// The result of a batching operation.
+pub enum BatchResult {
+    /// The `other` item was batched into `self`
+    Success,
+    /// `self` and `other` cannot be batched together
+    IncompatibleItems,
+}
+
+/// This system sorts the [`PhaseItem`]s of all [`RenderPhase`]s of this type.
 pub fn sort_phase_system<I: PhaseItem>(mut render_phases: Query<&mut RenderPhase<I>>) {
     for mut phase in &mut render_phases {
         phase.sort();
@@ -92,11 +265,9 @@ pub fn batch_phase_system<I: BatchedPhaseItem>(mut render_phases: Query<&mut Ren
 
 #[cfg(test)]
 mod tests {
-    use std::ops::Range;
-
-    use bevy_ecs::entity::Entity;
-
     use super::*;
+    use bevy_ecs::entity::Entity;
+    use std::ops::Range;
 
     #[test]
     fn batching() {
@@ -108,7 +279,7 @@ mod tests {
         impl PhaseItem for TestPhaseItem {
             type SortKey = ();
 
-            fn entity(&self) -> bevy_ecs::entity::Entity {
+            fn entity(&self) -> Entity {
                 self.entity
             }
 
@@ -119,11 +290,11 @@ mod tests {
             }
         }
         impl BatchedPhaseItem for TestPhaseItem {
-            fn batch_range(&self) -> &Option<std::ops::Range<u32>> {
+            fn batch_range(&self) -> &Option<Range<u32>> {
                 &self.batch_range
             }
 
-            fn batch_range_mut(&mut self) -> &mut Option<std::ops::Range<u32>> {
+            fn batch_range_mut(&mut self) -> &mut Option<Range<u32>> {
                 &mut self.batch_range
             }
         }
